Rename Match3 -> Board for extension classes.
