+++ Project/Assets/ML-Agents/Examples/Match3/Scripts/Match3Board.cs
@@ -1,4 +1,4 @@
-using Unity.MLAgents.Extensions.Match3;
+using Unity.MLAgents.Extensions.Board;
 using UnityEngine;
 
 namespace Unity.MLAgentsExamples
